Abstract inverter poller
To support non-SunSpec inverters in the future
