+++ package.json
@@ -12,7 +12,6 @@
         "start:debug": "node --inspect-brk=0.0.0.0:9229 dist/src/app.js",
         "dev": "npm run prebuild && tsx src/app.ts",
         "debug:sunspec-discovery": "tsx scripts/debugSunspecDiscovery.ts",
-        "debug:monitoring": "tsx scripts/debugMonitoring.ts",
         "cert:device-request": "tsx scripts/certDeviceRequest.ts",
         "cert:device-generate": "tsx scripts/certDeviceGenerate.ts",
         "generate:config-json-schema": "tsx scripts/generateConfigJsonSchema.ts",
